unify resources view models
